Rename tsinterval to date-interval
