+++ tests/unit/utils/date-interval-test.js
@@ -1,0 +1,10 @@
+import dateInterval from 'analytics-dashboard/utils/date-interval';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | date interval');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = dateInterval();
+  assert.ok(result);
+});
