--- tests/unit/utils/tsinterval-test.js
@@ -1,10 +1,0 @@
-import tsinterval from 'analytics-dashboard/utils/tsinterval';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | tsinterval');
-
-// Replace this with your real tests.
-test('it works', function(assert) {
-  let result = tsinterval();
-  assert.ok(result);
-});
